Work on

+++ src/world/world_constants.rs
@@ -1,4 +1,4 @@
-pub const CHUNK_SIZE: i32 = 16;
-pub const CHUNK_AREA: i32 = CHUNK_SIZE * CHUNK_SIZE;
-pub const CHUNK_VOLUME: i32 = CHUNK_SIZE * CHUNK_SIZE * CHUNK_SIZE;
-pub const WATER_LEVEL: i32 = 64;
+pub const CHUNK_SIZE: usize = 16;
+pub const CHUNK_AREA: usize = CHUNK_SIZE * CHUNK_SIZE;
+pub const CHUNK_VOLUME: usize = CHUNK_SIZE * CHUNK_SIZE * CHUNK_SIZE;
+pub const WATER_LEVEL: usize = 64;
